CodeFixes

+++ ShopTARge23/ShopTARge23.ApplicationServices/Services/FileServices.cs
@@ -25,59 +25,57 @@ namespace ShopTARge23.ApplicationServices.Services
         {
             if (dto.Files != null && dto.Files.Count > 0)
             {
-                if (!Directory.Exists(_webHost.ContentRootPath + "\\multipleFileUpload\\"))
-            if (!Directory.Exists(Path.Combine(_webHost.ContentRootPath, "multipleFileUpload")))
-            {
-                Directory.CreateDirectory(Path.Combine(_webHost.ContentRootPath, "multipleFileUpload"));
-            }
-
-            foreach (var file in dto.Files)
-            {
                 string uploadsFolder = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload");
-                string uniqueFileName = Guid.NewGuid().ToString() + "_" + file.FileName;
-                string filePath = Path.Combine(uploadsFolder, uniqueFileName);
 
-                using (var fileStream = new FileStream(filePath, FileMode.Create))
+                // Ensure the directory exists
+                if (!Directory.Exists(uploadsFolder))
                 {
-                    Directory.CreateDirectory(_webHost.ContentRootPath + "\\multipleFileUpload\\");
+                    Directory.CreateDirectory(uploadsFolder);
                 }
 
                 foreach (var file in dto.Files)
                 {
-                    string uploadsFolder = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload");
                     string uniqueFileName = Guid.NewGuid().ToString() + "_" + file.FileName;
                     string filePath = Path.Combine(uploadsFolder, uniqueFileName);
 
                     using (var fileStream = new FileStream(filePath, FileMode.Create))
                     {
-                        file.CopyTo(fileStream);
-                        FileToApi path = new FileToApi
-                        {
-                            Id = Guid.NewGuid(),
-                            ExistingFilePath = uniqueFileName,
-                            SpaceshipId = spaceship.Id
-                        };
-
-                        _context.FileToApis.AddAsync(path);
+                        file.CopyTo(fileStream); // Copy file to the directory
                     }
+
+                    // Create and add the new FileToApi entity
+                    var path = new FileToApi
+                    {
+                        Id = Guid.NewGuid(),
+                        ExistingFilePath = uniqueFileName,
+                        SpaceshipId = spaceship.Id
+                    };
+
                     _context.FileToApis.Add(path);
                 }
+
+                _context.SaveChanges(); // Save changes after all files are processed
             }
         }
 
         public async Task<FileToApi> RemoveImageFromApi(FileToApiDto dto)
         {
-            var imageId = await _context.FileToApis
+            var image = await _context.FileToApis
                 .FirstOrDefaultAsync(x => x.Id == dto.Id);
-            var filePath = _webHost.ContentRootPath + "\\multipleFileUpload\\"
-                + imageId.ExistingFilePath;
-            if (File.Exists(filePath))
+
+            if (image != null)
             {
-                File.Delete(filePath);
+                var filePath = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload", image.ExistingFilePath);
+                if (File.Exists(filePath))
+                {
+                    File.Delete(filePath); // Delete the file
+                }
+
+                _context.FileToApis.Remove(image);
+                await _context.SaveChangesAsync(); // Save changes
             }
-            _context.FileToApis.Remove(imageId);
-            await _context.SaveChangesAsync();
-            return null;
+
+            return image; // Return the removed image
         }
 
         public async Task<List<FileToApi>> RemoveImagesFromApi(FileToApiDto[] dtos)
@@ -87,16 +85,14 @@ namespace ShopTARge23.ApplicationServices.Services
             foreach (var dto in dtos)
             {
                 var image = await _context.FileToApis
-                    .FirstOrDefaultAsync(x => x.ExistingFilePath == dto.ExistingFilePath);
-                var filePath = _webHost.ContentRootPath + "\\multipleFileUpload\\" + imageId.ExistingFilePath;
-
+                    .FirstOrDefaultAsync(x => x.Id == dto.Id);
 
                 if (image != null)
                 {
                     var filePath = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload", image.ExistingFilePath);
                     if (File.Exists(filePath))
                     {
-                        File.Delete(filePath);
+                        File.Delete(filePath); // Delete the file
                     }
 
                     imagesToRemove.Add(image);
@@ -104,58 +100,63 @@ namespace ShopTARge23.ApplicationServices.Services
             }
 
             _context.FileToApis.RemoveRange(imagesToRemove);
-            await _context.SaveChangesAsync();
-
-            return imagesToRemove;
+            await _context.SaveChangesAsync(); // Save changes
+            return imagesToRemove; // Return the list of removed images
         }
 
         public void UploadFilesToData(KindergartenDto dto, Kindergarten domain)
         {
             if (dto.Files != null && dto.Files.Count > 0)
             {
+                string uploadsFolder = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload");
+                if (!Directory.Exists(uploadsFolder))
+                {
+                    Directory.CreateDirectory(uploadsFolder);
+                }
+
                 foreach (var image in dto.Files)
                 {
-                    string uploadsFolder = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload");
                     string uniqueFileName = Guid.NewGuid().ToString() + "_" + image.FileName;
                     string filePath = Path.Combine(uploadsFolder, uniqueFileName);
 
                     using (var fileStream = new FileStream(filePath, FileMode.Create))
                     {
-                        image.CopyTo(fileStream);
-
-                        FileToData files = new FileToData()
-                        {
-                            Id = Guid.NewGuid(),
-                            ImageTitles = image.FileName,
-                            ExistingFilePath = uniqueFileName, // Store the path instead of image data
-                            KindergartenId = domain.Id
-                        };
-
-                        _context.FileToDatas.Add(files);
+                        image.CopyTo(fileStream); // Copy file to the directory
                     }
+
+                    var files = new FileToData()
+                    {
+                        Id = Guid.NewGuid(),
+                        ImageTitles = image.FileName,
+                        ExistingFilePath = uniqueFileName, // Store the path instead of image data
+                        KindergartenId = domain.Id
+                    };
+
+                    _context.FileToDatas.Add(files); // Add to context
                 }
+
+                _context.SaveChanges(); // Save changes after all files are processed
             }
         }
 
         public async Task<FileToData> RemoveImageFromData(FileToDataDto dto)
         {
             var image = await _context.FileToDatas
-                .Where(x => x.Id == dto.Id)
-                .FirstOrDefaultAsync();
+                .FirstOrDefaultAsync(x => x.Id == dto.Id);
 
             if (image != null)
             {
                 var filePath = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload", image.ExistingFilePath);
                 if (File.Exists(filePath))
                 {
-                    File.Delete(filePath);
+                    File.Delete(filePath); // Delete the file
                 }
 
-                _context.FileToApis.Remove(imageId);
-                await _context.SaveChangesAsync();
+                _context.FileToDatas.Remove(image);
+                await _context.SaveChangesAsync(); // Save changes
             }
 
-            return null;
+            return image; // Return the removed image
         }
 
         public void UploadFilesToDatabase(RealEstateDto dto, RealEstate domain)
@@ -166,78 +167,91 @@ namespace ShopTARge23.ApplicationServices.Services
                 {
                     using (var target = new MemoryStream())
                     {
-                        FileToDatabase files = new FileToDatabase()
+                        image.CopyTo(target); // Copy file to memory stream
+
+                        var files = new FileToDatabase()
                         {
                             Id = Guid.NewGuid(),
                             ImageTitle = image.FileName,
-                            RealEstateId = domain.Id
+                            RealEstateId = domain.Id,
+                            ImageData = target.ToArray() // Store image data in byte array
                         };
-                        image.CopyTo(target);
-                        files.ImageData = target.ToArray();
 
-                        _context.FileToDatabases.Add(files);
+                        _context.FileToDatabases.Add(files); // Add to context
                     }
                 }
+
+                _context.SaveChanges(); // Save changes after all files are processed
             }
         }
+
         public async Task<FileToDatabase> RemoveImageFromDatabase(FileToDatabaseDto dto)
         {
             var image = await _context.FileToDatabases
+                .FirstOrDefaultAsync(x => x.Id == dto.Id);
 
-                .Where(x => x.Id == dto.Id)
-                .FirstOrDefaultAsync();
+            if (image != null)
+            {
+                _context.FileToDatabases.Remove(image);
+                await _context.SaveChangesAsync(); // Save changes
+            }
 
-            _context.FileToDatabases.Remove(image);
-            await _context.SaveChangesAsync();
-
-            return image;
+            return image; // Return the removed image
         }
 
-        public async Task<FileToDatabase> RemoveImagesFromDatabase(FileToDatabaseDto[] dtos)
+        public async Task<List<FileToDatabase>> RemoveImagesFromDatabase(FileToDatabaseDto[] dtos)
         {
+            List<FileToDatabase> imagesToRemove = new List<FileToDatabase>();
+
             foreach (var dto in dtos)
             {
                 var image = await _context.FileToDatabases
-                    .Where(x => x.Id == dto.Id)
-                    .FirstOrDefaultAsync();
+                    .FirstOrDefaultAsync(x => x.Id == dto.Id);
 
-                _context.FileToDatabases.Remove(image);
-                await _context.SaveChangesAsync();
-            }
-            return null;
-                _context.FileToDatas.Remove(image);
-                await _context.SaveChangesAsync();
+                if (image != null)
+                {
+                    imagesToRemove.Add(image); // Capture the image to be removed
+                }
             }
 
-            return image; // Consider returning null if the image is not found
+            if (imagesToRemove.Count > 0)
+            {
+                _context.FileToDatabases.RemoveRange(imagesToRemove);
+                await _context.SaveChangesAsync(); // Save changes
+            }
+
+            return imagesToRemove; // Return the list of removed images
         }
 
-        public async Task<FileToData> RemoveImagesFromData(FileToDataDto[] dtos)
+        public async Task<List<FileToData>> RemoveImagesFromData(FileToDataDto[] dtos)
         {
             List<FileToData> imagesToRemove = new List<FileToData>(); // Initialize a list to hold removed images
 
             foreach (var dataDto in dtos)
             {
+                // Retrieve the image based on the provided ID
                 var image = await _context.FileToDatas
-                    .Where(x => x.Id == dataDto.Id)
-                    .FirstOrDefaultAsync();
+                    .FirstOrDefaultAsync(x => x.Id == dataDto.Id);
 
+                // If the image exists, proceed to remove it
                 if (image != null)
                 {
+                    // Construct the file path for deletion
                     var filePath = Path.Combine(_webHost.ContentRootPath, "multipleFileUpload", image.ExistingFilePath);
                     if (File.Exists(filePath))
                     {
                         File.Delete(filePath); // Delete the physical file
                     }
 
-                    _context.FileToDatas.Remove(image);
-                    imagesToRemove.Add(image); // Capture the removed image
+                    _context.FileToDatas.Remove(image); // Remove the image from the context
+                    imagesToRemove.Add(image); // Add the removed image to the list
                 }
             }
 
-            await _context.SaveChangesAsync();
+            await _context.SaveChangesAsync(); // Save all changes to the database
 
-            return imagesToRemove.Count > 0 ? imagesToRemove.Last() : null; // Return the last removed image, or null if none were removed
+            return imagesToRemove; // Return the list of removed images
         }
+
     }
 }
